Created a separate MuonToSimAssociatorBase
Previously, code was getting a TrackAssociatorBase from the EventSetup
and casting it to a MuonAssociatorByHits in order to do muon association.
This change creates a separate MuonToSimAssociatorBase which provides
the correct interface. This new object can be obtained from the EventSetup
and is created by MuonAssociatorESProducer. Common code was moved to
MuonAssociatorByHitsHelper.
This change is needed as a step towards fixing the consumes problem
with TrackAssociatorBase.
