+++ SimMuon/MCTruth/plugins/MuonAssociatorESProducer.cc
@@ -6,6 +6,7 @@
 #include "FWCore/Framework/interface/ModuleFactory.h"
 
 #include "SimMuon/MCTruth/interface/MuonAssociatorByHits.h"
+#include "SimMuon/MCTruth/interface/MuonToSimAssociatorByHits.h"
 #include "SimTracker/Records/interface/TrackAssociatorRecord.h"
 
 class MuonAssociatorESProducer : public edm::ESProducer
@@ -15,6 +16,9 @@ public:
     MuonAssociatorESProducer(edm::ParameterSet const & config) : config_(config)
     {
         setWhatProduced(this, config.getParameter<std::string>("ComponentName"));
+
+        setWhatProduced(this, &MuonAssociatorESProducer::produceMuonAssociator,
+                        edm::es::Label(config.getParameter<std::string>("ComponentName")));
     }
 
     ~MuonAssociatorESProducer() {}
@@ -26,6 +30,13 @@ public:
         using namespace edm::es;
         boost::shared_ptr<TrackAssociatorBase> pMuonAssociatorByHits( new MuonAssociatorByHits(config_) );
         return pMuonAssociatorByHits;
+    }
+
+    boost::shared_ptr<MuonToSimAssociatorBase>
+    produceMuonAssociator(const TrackAssociatorRecord& record)
+    {
+      boost::shared_ptr<MuonToSimAssociatorBase> ret( new MuonToSimAssociatorByHits(config_) );
+      return ret;
     }
 
 private:
