+++ SimMuon/MCTruth/src/MuonToSimAssociatorBase.cc
@@ -1,0 +1,39 @@
+// -*- C++ -*-
+//
+// Package:     SimMuon/MCTruth
+// Class  :     MuonToSimAssociatorBase
+// 
+// Implementation:
+//     [Notes on implementation]
+//
+// Original Author:  Christopher Jones
+//         Created:  Tue, 06 Jan 2015 20:14:49 GMT
+//
+
+// system include files
+
+// user include files
+#include "SimMuon/MCTruth/interface/MuonToSimAssociatorBase.h"
+#include "FWCore/Utilities/interface/typelookup.h"
+
+
+//
+// constants, enums and typedefs
+//
+
+//
+// static data member definitions
+//
+
+//
+// constructors and destructor
+//
+MuonToSimAssociatorBase::MuonToSimAssociatorBase()
+{
+}
+
+MuonToSimAssociatorBase::~MuonToSimAssociatorBase()
+{
+}
+
+TYPELOOKUP_DATA_REG(MuonToSimAssociatorBase);
