+++ Validation/RecoMuon/src/RecoMuonValidator.cc
@@ -3,13 +3,14 @@
 #include "DataFormats/MuonReco/interface/MuonSelectors.h"
 #include "SimTracker/Records/interface/TrackAssociatorRecord.h"
 #include "SimTracker/TrackAssociation/interface/TrackAssociatorBase.h"
-#include "SimMuon/MCTruth/interface/MuonAssociatorByHits.h"
+#include "SimMuon/MCTruth/interface/MuonToSimAssociatorBase.h"
 
 #include "FWCore/ServiceRegistry/interface/Service.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
 //#include "DQMServices/Core/interface/DQMStore.h"
 #include "DQMServices/Core/interface/MonitorElement.h"
 
+#include "TrackingTools/TransientTrackingRecHit/interface/TransientTrackingRecHit.h"
 #include "TrackingTools/TransientTrack/interface/TransientTrack.h"
 #include "RecoMuon/TrackingTools/interface/MuonServiceProxy.h"
 
@@ -508,10 +509,10 @@ RecoMuonValidator::RecoMuonValidator(const edm::ParameterSet& pset):
 
   //type of track
   std::string trackType = pset.getParameter< std::string >("trackType");
-  if (trackType == "inner") trackType_ = MuonAssociatorByHits::InnerTk;
-  else if (trackType == "outer") trackType_ = MuonAssociatorByHits::OuterTk;
-  else if (trackType == "global") trackType_ = MuonAssociatorByHits::GlobalTk;
-  else if (trackType == "segments") trackType_ = MuonAssociatorByHits::Segments;
+  if (trackType == "inner") trackType_ = MuonToSimAssociatorBase::InnerTk;
+  else if (trackType == "outer") trackType_ = MuonToSimAssociatorBase::OuterTk;
+  else if (trackType == "global") trackType_ = MuonToSimAssociatorBase::GlobalTk;
+  else if (trackType == "segments") trackType_ = MuonToSimAssociatorBase::Segments;
   else throw cms::Exception("Configuration") << "Track type '" << trackType << "' not supported.\n";
 
 //  seedPropagatorName_ = pset.getParameter<string>("SeedPropagator");
@@ -616,9 +617,9 @@ void RecoMuonValidator::dqmBeginRun(const edm::Run& , const EventSetup& eventSet
   if ( theMuonService ) theMuonService->update(eventSetup);
 
   if ( doAssoc_ ) {
-    edm::ESHandle<TrackAssociatorBase> associatorBase;
+    edm::ESHandle<MuonToSimAssociatorBase> associatorBase;
     eventSetup.get<TrackAssociatorRecord>().get(muAssocLabel_.label(), associatorBase);
-    assoByHits = dynamic_cast<const MuonAssociatorByHits *>(associatorBase.product());
+    assoByHits = associatorBase.product();
     if (assoByHits == 0) throw cms::Exception("Configuration") << "The Track Associator with label '" << muAssocLabel_.label() << "' is not a MuonAssociatorByHits.\n";
     }
 
@@ -692,8 +693,8 @@ void RecoMuonValidator::analyze(const Event& event, const EventSetup& eventSetup
   muonME_->hNSim_->Fill(nSim);
   muonME_->hNMuon_->Fill(muonColl.size());
 
-  MuonAssociatorByHits::MuonToSimCollection muonToSimColl;
-  MuonAssociatorByHits::SimToMuonCollection simToMuonColl;
+  MuonToSimAssociatorBase::MuonToSimCollection muonToSimColl;
+  MuonToSimAssociatorBase::SimToMuonCollection simToMuonColl;
 
 
   if ( doAssoc_ ) {
